auto_test/requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
altgraph==0.17
apipkg==1.5
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coverage==5.3.1
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
demjson==2.2.4
enum34==1.1.10
execnet==1.7.1
future==0.18.2
idna==2.9
importlib-metadata==1.5.0
iniconfig==1.1.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jsonpath==0.82
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.3.2
more-itertools==8.2.0
namedlist==1.8
netifaces==0.10.9
numpy==1.19.4
packaging==20.1
paramiko==2.7.1
parso==0.7.1
pefile==2019.4.18
pickleshare==0.7.5
pika==1.1.0
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.8
py==1.10.0
pycparser==2.19
Pygments==2.7.2
PyInstaller==3.6
PyNaCl==1.3.0
pyparsing==2.4.6
pytest==4.5.0
pytest-assume==2.4.2
pytest-base-url==1.4.2
pytest-cov==2.10.1
pytest-forked==1.3.0
pytest-html==1.22.1.dev11+g4f729c1
pytest-instafail==0.4.2
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-reportlog==0.1.2
pytest-rerunfailures==9.1.1
pytest-sugar==0.9.4
pytest-xdist==2.2.0
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
PyX==0.15
requests==2.23.0
scapy==2.4.3
selenium==3.141.0
six==1.14.0
termcolor==1.1.0
toml==0.10.2
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.8
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==3.0.0
